visis_bot/requirements.txt
aiogram==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai>=2.0.0